experimental_resumeDPMUnc - Resume run of Dirichlet Process Mixture Modeller taking uncertainty of data points into account

experimental_resumeDPMUnc - Resume run of Dirichlet Process Mixture Modeller taking uncertainty of data points into account

Usage

experimental_resumeDPMUnc(
  obsData,
  obsVars,
  saveFileDir,
  seed,
  kappa0,
  alpha0,
  beta0,
  K = floor(nrow(obsData)/2),
  nIts = 1e+05,
  thinningFreq = 10,
  saveClusterParams = TRUE,
  saveLatentObs = FALSE,
  quiet = TRUE,
  scaleData = FALSE
)

Arguments

obsData

The observed data in matrix form (n observations x p variables)

obsVars

The observed variances of the data (n observations x p variables)

saveFileDir

Directory where all output will be saved, and where existing output should be found

seed

Seed for random number generator, to make the function deterministic.

kappa0

Hyperparameter for Normal-Gamma prior on cluster parameters

alpha0

Hyperparameter for Normal-Gamma prior on cluster parameters

beta0

Hyperparameter for Normal-Gamma prior on cluster parameters

K

Initial number of clusters.

nIts

Total number of iterations to run. The user should check they are happy that the model has converged before using any of the results.

thinningFreq

Controls how many samples are saved. E.g. a value of 10 means every 10th sample will be saved.

saveClusterParams

Boolean, determining whether the cluster parameters (mean and variance of every cluster) for every saved iteration should be saved in a file or not. Both cluster parameters and latent observations take up more space than other saved variables. The files clusterVars.csv and clusterMeans.csv will be created in either case, but will be left empty if saveClusterParams is FALSE.

saveLatentObs

Boolean, determining whether the latent observations (underlying true observations) for every saved iteration should be saved in a file or not. Both cluster parameters and latent observations take up more space than other saved variables. The file latentObservations.csv will be created in either case, but will be left empty if saveLatentObs is FALSE.

quiet

Boolean. If FALSE, information will be printed to the terminal including current iteration, current value of K and number of items per cluster.

scaleData

Boolean. If TRUE, data will be scaled so that the variance of every variable (column) in obsData is 1 (and obsVars will be scaled to fit this rescaling). Else, the raw data will be used.

Examples

n = 50; d = 5; k = 4
classes = sample(1:k, n, replace=TRUE)
group_means = matrix(rep(classes, d), n, d)
true_means = group_means + matrix(rnorm(n*d, sd=0.1),n,d)
obsVars = matrix(rchisq(n*d,1), n, d)
obsData = matrix(rnorm(n*d, mean=as.vector(true_means), sd=sqrt(as.vector(obsVars))), n, d)
# The hyperparameters should be carefully checked against the data.
alpha0 = 2; beta0 = 0.2 * mean(apply(obsData, 2, var)); kappa0 = 0.5
#DPMUnc(obsData, obsVars, "test_output", 1234,
#       kappa0=kappa0, alpha0=alpha0, beta0=beta0, nIts=10000)
#experimental_resumeDPMUnc(obsData, obsVars, "test_output", 1234,
#                          kappa0=kappa0, alpha0=alpha0, beta0=beta0, nIts=100000)